+++ extra/dependencyLibs/inputmask.dependencyLib.jqlite.js
@@ -6,7 +6,7 @@
  */
 (function (factory) {
 	if (typeof define === "function" && define.amd) {
-		define(["jqlite"], factory);
+		define("inputmask.dependencyLib", ["jqlite"], factory);
 	} else if (typeof exports === "object") {
 		module.exports = factory(require("jqlite"));
 	} else {