+++ extra/dependencyLibs/inputmask.dependencyLib.js
@@ -6,7 +6,7 @@
  */
 (function (factory) {
 	if (typeof define === "function" && define.amd) {
-		define(factory);
+		define("inputmask.dependencyLib", factory);
 	} else if (typeof exports === "object") {
 		module.exports = factory();
 	} else {