+++ js/inputmask.dependencyLib.js
@@ -1,6 +1,6 @@
 (function (factory) {
 	if (typeof define === "function" && define.amd) {
-		define(["jquery"], factory);
+		define("inputmask.dependencyLib", ["jquery"], factory);
 	} else if (typeof exports === "object") {
 		module.exports = factory(require("jquery"));
 	} else {